[travis-ci] Add Travis CI support
This commit adds support for Travis CI to the codebase. My thoughts are
that by automating some of the checking features, I can make sure that
my pull requests are validated by someone (or something) other than
myself. The configuration has  mostly been stolen from the Regent
configuration, but should be sufficient for this use case.