[eo12958] Add tests for compliance with EO 12958
Executive Order 12958 Section 1.5 requires classifiers to mark the
reason a document was classified. Markings are simple; they're just the
section number "1.5" with the paragraph number in parenthesis "1.5(a)".
This changes a little bit every time the EO is rescinded or amended, but
not enough to justify a more complicated representation of the data at
this time.