[issue-18] Fix the (U) + (TS//REL) = (TS) bug
This commit fixes the bug reported in Ticket #18. This fix corrects the
issue by checking the classification levels of two Classification
objects prior to merge, and then caching any relevant REL TO elements
before they are merged together. If there are any cached REL TO
elements, then they are mixed back into the classification after the
merge is complete.

This implementation helps prevent the Dissemination class from needing
to be aware of the classification levels of each classification outside
of the class, and sets up a precedent moving forward for where these
kinds of rule-oddities should be handled.